Fix: backend integration
