+++ backend/app/api/main.py
@@ -4,5 +4,4 @@ from fastapi import APIRouter
 from app.api.routes import login, users
 
 api_router = APIRouter()
-api_router.include_router(login.router, tags=["login"])
 api_router.include_router(users.router, prefix="/users", tags=["users"])
